arch/x86/mm.c: do not set readonly flag when UBOOT binary provided
When u-boot binary was provided - it places inside .text section. This
is needed for the memory to have execute flag. When systems jump to
the u-boot it start using stack and data, included in the u-boot. For
this purpose there should be write access to the memory.
That's why set_readonly parameters skips when UBOOT_BIN was configured.

Signed-off-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>